Auto
TASK FORCE ANNOUNCEMENT
Let it be known that if going to "school" conflicts with a scheduled event, you are not permitted to go to school. Your loyalty to the Mobile Task Force is non negotiable. There is no need for school when you are already enrolled in a Task Force academy.
React if you understand.